shipping_label_py/requirements.txt
Pillow>=10.0.0
PyMuPDF>=1.23.0
reportlab>=4.0.0
Flask>=3.0.0
